// src/vs/platform/agentHost/node/codex/protocol/generated/v2/ConfigReadResponse.ts
// AUTOGENERATED - do not edit.
// Source: `codex app-server generate-ts --experimental`
// Generated from @openai/codex 0.153.0, licensed Apache-2.0.
// Modified by VS Code to rewrite imports and apply repository formatting.
// See README.md in this directory for provenance and licensing notes.
// Regenerate: npm run codex:gen-protocol
// Source file: v2/ConfigReadResponse.ts

import type { Config } from "./Config.js";
import type { ConfigLayer } from "./ConfigLayer.js";
import type { ConfigLayerMetadata } from "./ConfigLayerMetadata.js";

export type ConfigReadResponse = { config: Config, origins: { [key in string]?: ConfigLayerMetadata }, layers: Array<ConfigLayer> | null, };